Loosen the assumption of MDA.
